RegisterProtobufFormatters() => AddProtoBufNet() - mimic code from MvcNewtonsoftJsonOptionsExtensions etc
