+++ src/protobuf-net.TestWeb/Startup.cs
@@ -1,13 +1,8 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using ProtoBuf;
 
 namespace protobuf_net.TestWeb
 {
@@ -23,11 +18,7 @@ namespace protobuf_net.TestWeb
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddMvc(options =>
-            {
-                options.RegisterProtobufFormatters();
-            });
-            services.AddControllersWithViews();
+            services.AddControllersWithViews().AddProtoBufNet();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
